Feat: ISSUE 도메인 Response 수정 / 토큰 활용해서 인증 인가

+++ src/main/kotlin/com/arij/ajir/domain/issue/service/IssueService.kt
@@ -1,9 +1,9 @@
 package com.arij.ajir.domain.issue.service
 
+import com.arij.ajir.domain.comment.model.toResponse
 import com.arij.ajir.common.exception.ModelNotFoundException
-import com.arij.ajir.domain.issue.dto.IssueCreateRequest
-import com.arij.ajir.domain.issue.dto.IssueResponse
-import com.arij.ajir.domain.issue.dto.IssueUpdateRequest
+import com.arij.ajir.domain.comment.repository.CommentRepository
+import com.arij.ajir.domain.issue.dto.*
 import com.arij.ajir.domain.issue.model.Issue
 import com.arij.ajir.domain.issue.model.Priority
 import com.arij.ajir.domain.issue.repository.IssueRepository
@@ -11,7 +11,6 @@ import com.arij.ajir.domain.member.model.Member
 import com.arij.ajir.domain.member.repository.MemberRepository
 import com.arij.ajir.domain.team.repository.TeamRepository
 import org.springframework.data.domain.Sort
-import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -20,8 +19,7 @@ class IssueService(
     private val issueRepository: IssueRepository,
     private val memberRepository: MemberRepository,
     private val teamRepository: TeamRepository,
-    // todo(private val memberRepository: MemberRepository,
-    // todo(private val teamRepository: TeamRepository,
+    private val commentRepository: CommentRepository,
 ) {
     fun getAllIssues(
         topic: String,
@@ -33,48 +31,61 @@ class IssueService(
         return null
     }
 
-    fun getIssueById(id: Long): IssueResponse {
+    fun getIssueById(id: Long): IssueResponseWithCommentResponse {
         val issue = issueRepository.findIssueByIdAndDeletedIsFalse(id)
             .orElseThrow() { IllegalStateException("Issue not found") }
+        val comment = commentRepository.findAllByIssueId(id).map { it.toResponse() }
 
-        return issue.toResponse()
+        return issue.toResponseWithCommentResponse(comment)
     }
 
     @Transactional
-    fun createIssue(request: IssueCreateRequest, email: String): IssueResponse {
+    fun createIssue(request: IssueCreateRequest, email: String): IssueIdResponse {
         val member: Member = memberRepository.findByEmail(email) ?: throw ModelNotFoundException("Member", email)
 
         val issue = Issue.createIssue(request, member, member.team!!)
 
-        return issueRepository.save(issue).toResponse()
+        return issueRepository.save(issue).toIdResponse()
     }
 
     @Transactional
-    fun updateIssue(issueId: Long, request: IssueUpdateRequest): IssueResponse {
+    fun updateIssue(issueId: Long, request: IssueUpdateRequest, email: String) {
         val (title, content, category) = request
+        val member: Member = memberRepository.findByEmail(email) ?: throw ModelNotFoundException("Member", email)
         val issue = issueRepository.findIssueByIdAndDeletedIsFalse(issueId)
             .orElseThrow() { IllegalStateException("Issue not found") }
+
+        if (issue.member != member && issue.team != member.team) {
+            throw IllegalStateException("member and team not same")
+        }
 
         issue.title = title
         issue.content = content
         issue.category = category
-
-        return issue.toResponse()
     }
 
     @Transactional
-    fun updatePriority(issueId:Long, newPriority: Priority):IssueResponse {
+    fun updatePriority(issueId: Long, newPriority: Priority, email: String) {
+        val member: Member = memberRepository.findByEmail(email) ?: throw ModelNotFoundException("Member", email)
         val issue = issueRepository.findIssueByIdAndDeletedIsFalse(issueId)
             .orElseThrow() { IllegalStateException("Issue not found") }
-        issue.priority = newPriority
 
-        return issueRepository.save(issue).toResponse()
+        if (issue.team != member.team) {
+            throw IllegalStateException("team not same")
+        }
+
+        issue.priority = newPriority
     }
 
     @Transactional
-    fun deleteIssue(id: Long) {
+    fun deleteIssue(id: Long, email: String) {
+        val member: Member = memberRepository.findByEmail(email) ?: throw ModelNotFoundException("Member", email)
         val issue = issueRepository.findIssueByIdAndDeletedIsFalse(id)
             .orElseThrow() { IllegalStateException("Issue not found") }
+
+        if (issue.team != member.team) {
+            throw IllegalStateException("team not same")
+        }
 
         issue.delete()
         issueRepository.save(issue)
